+++ ChangeLog.md
@@ -3,6 +3,7 @@ ChangeLog
 
 Version 3.0
 -----------
+- (B.10) Added all necessary getters on the controller
 - (B.9) Created game controller, there is only model object creation implemented in it
 - (RC.1) Refactored all *dead* references to *knock out* in order to follow the spirit that a 0 HP
          character could be available for future combats.